CMSG_BATTLEFIELD_LIST structure update

+++ src/server/game/Server/Protocol/Handlers/BattleGroundHandler.cpp
@@ -316,7 +316,7 @@ void WorldSession::HandleBattlefieldListOpcode(WorldPacket &recv_data)
 {
     sLog.outDebug("WORLD: Recvd CMSG_BATTLEFIELD_LIST Message");
 
-    uint32 bgTypeId;
+    uint8 bgTypeId;
     recv_data >> bgTypeId;                                  // id from DBC
 
     WorldPacket data;
